requirements.txt
Django==1.8.2
djangorestframework==3.2.4
django-compressor==1.5
django-debug-toolbar==1.3.2
django-extensions==1.5.5
ipython==3.2.0
jsonfield==1.0.3
six==1.9.0
sqlparse==0.1.15
wsgiref==0.1.2
coverage>=3.7.1
coveralls>=0.5
